web_app_system/Data_Infrastructure/requirements.txt
pandas>=2.0.2
requests>=2.31.0
beautifulsoup4>=4.12.2
google-cloud-bigquery>=3.11.3
